requirements.txt
jinja2 == 3.1.6
kubernetes == 34.1.0
pandas == 2.3.3
pyyaml == 6.0.3
pydantic == 2.12.5
scikit-learn == 1.7.2
toml == 0.10.2
torch == 2.9.1
ruamel.yaml == 0.19.1
numpy == 2.3.5
scipy == 1.16.3
requests == 2.32.5
websocket-client == 1.9.0
